Extract compression module from serde to simplify re-use (#2836)
This makes the compression and de-compression easier to use again in a
second serialization approach (e.g. Protobuf.) Resolves #2834.
