+++ syft/serde/__init__.py
@@ -2,5 +2,5 @@ from syft.serde.serde import *
 from syft.serde.torch_serde import *
 from syft.serde.serde import _simplify
 from syft.serde.serde import _detail
-from syft.serde.serde import _compress
-from syft.serde.serde import _decompress
+from syft.serde.compression import _compress
+from syft.serde.compression import _decompress
