+++ syft/serde/serde.py
@@ -35,15 +35,11 @@ from collections import OrderedDict
 from typing import Callable
 
 import inspect
-import lz4
-from lz4 import (  # noqa: F401
-    frame,
-)  # needed as otherwise we will get: module 'lz4' has no attribute 'frame'
 import msgpack
-import zstd
 
 import syft
 from syft import dependency_check
+
 from syft.federated.train_config import TrainConfig
 from syft.frameworks.torch.tensors.decorators.logging import LoggingTensor
 from syft.frameworks.torch.tensors.interpreters.precision import FixedPrecisionTensor
@@ -71,11 +67,11 @@ from syft.messaging.message import GetShapeMessage
 from syft.messaging.message import ForceObjectDeleteMessage
 from syft.messaging.message import SearchMessage
 from syft.messaging.message import PlanCommandMessage
+from syft.serde import compression
 from syft.serde.native_serde import MAP_NATIVE_SIMPLIFIERS_AND_DETAILERS
 from syft.workers.abstract import AbstractWorker
 from syft.workers.base import BaseWorker
 
-from syft.exceptions import CompressionNotFoundException
 from syft.exceptions import GetNotPermittedError
 from syft.exceptions import ResponseSignatureError
 
@@ -143,16 +139,6 @@ OBJ_FORCE_FULL_SIMPLIFIER_AND_DETAILERS = [BaseWorker]
 # NOTE: serialization constants for these objects need to be defined in `proto.json` file
 # in https://github.com/OpenMined/proto
 EXCEPTION_SIMPLIFIER_AND_DETAILERS = [GetNotPermittedError, ResponseSignatureError]
-
-# COMPRESSION SCHEME INT CODES
-NO_COMPRESSION = 40
-LZ4 = 41
-ZSTD = 42
-scheme_to_bytes = {
-    NO_COMPRESSION: NO_COMPRESSION.to_bytes(1, byteorder="big"),
-    LZ4: LZ4.to_bytes(1, byteorder="big"),
-    ZSTD: ZSTD.to_bytes(1, byteorder="big"),
-}
 
 ## SECTION: High Level Simplification Router
 def _force_full_simplify(worker: AbstractWorker, obj: object) -> object:
@@ -300,7 +286,7 @@ def _serialize_msgpack_binary(
     # otherwise we output the compressed stream with header set to '1'
     # even if compressed flag is set to false by the caller we
     # output the input stream as it is with header set to '0'
-    return _compress(binary)
+    return compression._compress(binary)
 
 
 def _serialize_msgpack(
@@ -344,7 +330,7 @@ def _deserialize_msgpack_binary(binary: bin, worker: AbstractWorker = None) -> o
         worker = syft.framework.hook.local_worker
 
     # 1) Decompress the binary if needed
-    binary = _decompress(binary)
+    binary = compression._decompress(binary)
 
     # 2) Deserialize
     # This function converts the binary into the appropriate python
@@ -434,113 +420,6 @@ def deserialize(
         object: the deserialized form of the binary input.
     """
     return strategy(binary, worker)
-
-
-## SECTION: chosen Compression Algorithm
-
-
-def _apply_compress_scheme(decompressed_input_bin) -> tuple:
-    """
-    Apply the selected compression scheme.
-    By default is used LZ4
-
-    Args:
-        decompressed_input_bin: the binary to be compressed
-    """
-    return apply_lz4_compression(decompressed_input_bin)
-
-
-def apply_lz4_compression(decompressed_input_bin) -> tuple:
-    """
-    Apply LZ4 compression to the input
-
-    Args:
-        decompressed_input_bin: the binary to be compressed
-
-    Returns:
-        a tuple (compressed_result, LZ4)
-    """
-    return lz4.frame.compress(decompressed_input_bin), LZ4
-
-
-def apply_zstd_compression(decompressed_input_bin) -> tuple:
-    """
-    Apply ZSTD compression to the input
-
-    Args:
-        decompressed_input_bin: the binary to be compressed
-
-    Returns:
-        a tuple (compressed_result, ZSTD)
-    """
-
-    return zstd.compress(decompressed_input_bin), ZSTD
-
-
-def apply_no_compression(decompressed_input_bin) -> tuple:
-    """
-    No compression is applied to the input
-
-    Args:
-        decompressed_input_bin: the binary
-
-    Returns:
-        a tuple (the binary, LZ4)
-    """
-
-    return decompressed_input_bin, NO_COMPRESSION
-
-
-def _compress(decompressed_input_bin: bin) -> bin:
-    """
-    This function compresses a binary using the function _apply_compress_scheme
-    if the input has been already compressed in some step, it will return it as it is
-
-    Args:
-        decompressed_input_bin (bin): binary to be compressed
-
-    Returns:
-        bin: a compressed binary
-
-    """
-    compress_stream, compress_scheme = _apply_compress_scheme(decompressed_input_bin)
-    try:
-        z = scheme_to_bytes[compress_scheme] + compress_stream
-        return z
-    except KeyError:
-        raise CompressionNotFoundException(
-            f"Compression scheme not found for compression code: {str(compress_scheme)}"
-        )
-
-
-def _decompress(binary: bin) -> bin:
-    """
-    This function decompresses a binary using the scheme defined in the first byte of the input
-
-    Args:
-        binary (bin): a compressed binary
-
-    Returns:
-        bin: decompressed binary
-
-    """
-
-    # check the 1-byte header to check the compression scheme used
-    compress_scheme = binary[0]
-
-    # remove the 1-byte header from the input stream
-    binary = binary[1:]
-    # 1)  Decompress or return the original stream
-    if compress_scheme == LZ4:
-        return lz4.frame.decompress(binary)
-    elif compress_scheme == ZSTD:
-        return zstd.decompress(binary)
-    elif compress_scheme == NO_COMPRESSION:
-        return binary
-    else:
-        raise CompressionNotFoundException(
-            f"Compression scheme not found for compression code: {str(compress_scheme)}"
-        )
 
 
 def _simplify(worker: AbstractWorker, obj: object, **kwargs) -> object:
